GenericPopupToolbarController: Support replacing the main command
If a second arg is passed in the "Value" property, it will be
treated as boolean. If true - the button will be
ToolBoxItemBits::DROPDOWN, and will keep replacing the main
command with the last selected one from the dropdown. It will
also respond to status updates of that command, currently for
enabled/disabled and boolean (true treated as pressed) states.

Change-Id: I09a5c20e6d2a010867037754f036096246749ec4