Avoid further crash with empty Chart document
following up on 6ab00804ceed30ccc4fdad987dca05d7657c2359 "Avoid crash with empty
Chart document", when you then close the Writer doc (discarding the changes), it
crashes in a chart::sidebar::ChartSidebarModifyListener::disposing call whose
mpParent (that was a ChartElementsPanel) is already destroyed.  The problem is
that the ChartElementsPanel ctor creates an mxListener pointing to this, then
calls Initialize, which calls xBroadcaster->addModifyListener(mxListener) and
then, during updateData, throws a RuntimeException when xDiagram is null in
getChartType, causing the ChartElementsPanel object to immediately be destroyed,
but the listener still registered.  This apparently needs fixing independently
of the immediate problem.  (The immediate problem is a regression introduced
with a835af16be70c3310c94bd0babd2e679a4dcc13f "hide grid and axes options in
chart elements if not supported, tdf#94297".)

Change-Id: I2fe4f2a59600664f6210456c65e033f9453f4cc9