tdf#101814 sw: MSWord export: don't override first-page with heuristics
Revert commit ed8e9fae7c838bf76902bc79709ee08f2024406a, and do the
opposite.  The added "titlePage" check makes no sense, because in that
case there is already an explicit first-page style in the PageDesc, and
overriding that cannot be right.

The test added in that commit still succeeds, which suggests that the
root cause of that bug was elsewhere and was fixed in the meantime.

Change-Id: I9b21b6dc115be28cb0ee3360615c538d4cfee3cb