tdf#103514: Always ignore bitmap fonts on Windows
We should never see any bitmap fonts now, so all the Courier stuff can
go away.

Since we never list the bitmap fonts in the first place, the old code
for replacing some of them with scalable equivalents never kicks in. We
instead implement PreMatchFontSubstitution hook for the same effect.

Change-Id: I54f3c068bcd64554ed3498f4ba5b0f76bbd38589
Reviewed-on: https://gerrit.libreoffice.org/30345
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>
Tested-by: Michael Meeks <michael.meeks@collabora.com>