tdf#103624 Avoid double init of controls in SidebarToolBox
Controls that created by the service manager are already
initialized by it. At least in the case of SaveToolbarController
such double init creates a problem, as SaveToolbarController
will add itself as a modify listener twice, but remove only
once in dispose method. Probably SaveToolbarController should
be smarter about this, for both not adding a listener twice,
and also not attempt to work after being disposed - but that
doesn't make the behavior of SidebarToolBox correct.

Change-Id: I3f0fcdff31c65198a2233c969d37a0015712fff5