sc-perf: eliminate unnecessary loops in ScTable::UpdateSelectionFunction()
That looped unconditionally over all columns of a sheet just to let
ScColumn::UpdateSelectionFunction() costly (by creating empty spans) decide
that it doesn't have to do anything. This for *every* cell cursor movement or
switching sheets et al. Instead, use the ScMarkData area to narrow down the
range beforehand, which when travelling with the cell cursor is just one
column.

Change-Id: Ic60928d07bc6cec4f6d8491ab30b99d7b20b8490