Add fill-column: 100 to the Emacs mode line
Vim users: No reason to panic. This does not mean that Emacs users
would suddenly start to split existing (or new) code lines always
before column 100. The default value for fill-column is 70, and that
hasn't forced code lines edited in Emacs to be shorter than that
earlier either.

The primary intent of using a fill-column of 100 (instead of the
default 70) is that when you edit some long (multi-line) comment
block, and you want to reformat ("fill", "reflow") that comment (using
the fill-paragraph command, bound to Meta-Q), lines will be filled up
to column 100, and not just 70, which in most cases would look quite
short.

Unless I am strongly advised not to, I will start adding this to the
mode lines in source files if I remember, as I happen to edit some
comment block in them.

Change-Id: Icfb93dbb22b2db7190fdc9c8ee9518d08e73c7a8