tdf#88827 - double-thin border: MINGAPWIDTH is a valid width
double-thin borders are available in the UI starting from 0.5pt.
The actual minumum (as seen in a round-trip), is 1.10pt.
(Each thin line is ~ .50pt, the gap is ~ .05pt, and then some
approximations and rounding show it as 1.10 - at least that is how I
understood it). 1.15pt is the first point at which the gap is larger
than the minimum - and double_thins with a minimum gap were considered
invalid, and thus were not imported.

With this fix, double-thin borders created with a size less than 1.15pt
are valid and visible on import.

Change-Id: I6da2a40d13ed83281de403b22b3acbea4288ac60
Reviewed-on: https://gerrit.libreoffice.org/30857
Reviewed-by: Justin Luth <justin_luth@sil.org>
Tested-by: Justin Luth <justin_luth@sil.org>