Resolves: tdf#103914 if menu no longer has focus, leave focus where it is
In the normal course of events, the menu, or its children, has focus when its
popped down, in this case continue to restored the focus to the previous focus
window which had it when the menu appeared.

If some other non-child window of the menu has focus as popdown time, leave
the focus where it is.

Change-Id: Ia860f90350653ad4d8056738dacbc434fb364989