tdf#96475 PutFormulaCell: any other cell than formula is utter nonsense
This is called only if a cell has the table:formula attribute and that *is* a
formula cell and nothing else. In fact in ODFF the initial leading '=' is not
mandatory, so attempting to set a different cell type if it is not present is
wrong.

Commit eb01b53d06cf7c607d4c0817b67d4a21dfae6fe7 introduced that, which at that
time may have been necessary, but doubtful..

Additionally, ScFormulaCell::CompileXML() that tries to group formulas had to
be adpated to not rely on the presence of a leading '='. Luckily there was an
assert..

These changes enable loading of "error cell" formulas that were stored without
a leading '=' if they originated from a paste special with only values, which
maintains an error result as error formula.

Change-Id: I43394de108066a24b792eec958b19f51f990403b